+++ TTOD2D/src/source/SkillController.java
@@ -95,14 +95,14 @@ public class SkillController {
             case 2:
                 skillToDefine.setName("Throw Scrap");
                 skillToDefine.setManaCost(10);
-                skillToDefine.setPotency(1.5);
+                skillToDefine.setPotency((float) 1.5);
                 skillToDefine.setUnlockedAt(2);
                 Player.getInstance().getSkills().add(skillToDefine);
                 break;
             case 4:
                 skillToDefine.setName("Double Strike");
                 skillToDefine.setManaCost(20);
-                skillToDefine.setPotency(2.5);
+                skillToDefine.setPotency((float) 2.5);
                 skillToDefine.setUnlockedAt(4);
                 Player.getInstance().getSkills().add(skillToDefine);
                 break;
@@ -118,28 +118,28 @@ public class SkillController {
                     case Human:
                         skillToDefine.setName("Ice Blast");
                         skillToDefine.setManaCost(100);
-                        skillToDefine.setPotency(3.5);
+                        skillToDefine.setPotency((float) 3.5);
                         skillToDefine.setUnlockedAt(8);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Dragonpeeps:
                         skillToDefine.setName("Fire Breath");
                         skillToDefine.setManaCost(100);
-                        skillToDefine.setPotency(3.5);
+                        skillToDefine.setPotency((float) 3.5);
                         skillToDefine.setUnlockedAt(8);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Big_Daddys:
                         skillToDefine.setName("Water Bullet");
                         skillToDefine.setManaCost(100);
-                        skillToDefine.setPotency(3.5);
+                        skillToDefine.setPotency((float) 3.5);
                         skillToDefine.setUnlockedAt(8);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Dwarf:
                         skillToDefine.setName("Quake");
                         skillToDefine.setManaCost(100);
-                        skillToDefine.setPotency(3.5);
+                        skillToDefine.setPotency((float) 3.5);
                         skillToDefine.setUnlockedAt(8);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
@@ -154,7 +154,7 @@ public class SkillController {
             case 12:
                 skillToDefine.setName("Obscene Gesture");
                 skillToDefine.setManaCost(50);
-                skillToDefine.setPotency(4.5);
+                skillToDefine.setPotency((float) 4.5);
                 skillToDefine.setUnlockedAt(12);
                 Player.getInstance().getSkills().add(skillToDefine);
                 break;
@@ -170,28 +170,28 @@ public class SkillController {
                     case Human:
                         skillToDefine.setName("Nature's Wrath");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(6.5);
+                        skillToDefine.setPotency((float) 6.5);
                         skillToDefine.setUnlockedAt(16);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Dragonpeeps:
                         skillToDefine.setName("Wingbeat");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(6.5);
+                        skillToDefine.setPotency((float) 6.5);
                         skillToDefine.setUnlockedAt(16);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Big_Daddys:
                         skillToDefine.setName("Iron Breacker");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(6.5);
+                        skillToDefine.setPotency((float) 6.5);
                         skillToDefine.setUnlockedAt(16);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Dwarf:
                         skillToDefine.setName("Shadow Orb");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(6.5);
+                        skillToDefine.setPotency((float) 6.5);
                         skillToDefine.setUnlockedAt(16);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
@@ -206,7 +206,7 @@ public class SkillController {
             case 20:
                 skillToDefine.setName("Reaper");
                 skillToDefine.setManaCost(120);
-                skillToDefine.setPotency(7.5);
+                skillToDefine.setPotency((float) 7.5);
                 skillToDefine.setUnlockedAt(20);
                 Player.getInstance().getSkills().add(skillToDefine);
                 break;
@@ -222,35 +222,35 @@ public class SkillController {
                     case Human:
                         skillToDefine.setName("Permafrost");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(8.5);
+                        skillToDefine.setPotency((float) 8.5);
                         skillToDefine.setUnlockedAt(24);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Dragonpeeps:
                         skillToDefine.setName("Firestorm");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(8.5);
+                        skillToDefine.setPotency((float) 8.5);
                         skillToDefine.setUnlockedAt(24);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Big_Daddys:
                         skillToDefine.setName("Hydraulic Impact");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(8.5);
+                        skillToDefine.setPotency((float) 8.5);
                         skillToDefine.setUnlockedAt(24);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                     case Dwarf:
                         skillToDefine.setName("Netherworld");
                         skillToDefine.setManaCost(250);
-                        skillToDefine.setPotency(8.5);
+                        skillToDefine.setPotency((float) 8.5);
                         skillToDefine.setUnlockedAt(24);
                         Player.getInstance().getSkills().add(skillToDefine);
                         break;
                 }
                 skillToDefine2.setName("Overdrive");
                 skillToDefine2.setManaCost(100);
-                skillToDefine2.setPotency(8.5);
+                skillToDefine2.setPotency((float) 8.5);
                 skillToDefine2.setUnlockedAt(24);
                 Player.getInstance().getSkills().add(skillToDefine2);
                 break;
